pyproject.toml
[tool.poetry]
name = "pitcher-analyzer"
version = "0.1.0"
description = "Baseball pitcher mechanics analysis tool using Vertex AI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<4.0"
google-cloud-storage = "^2.8.0"
opencv-python-headless = "^4.7.0"
numpy = "^1.24.0"
google-cloud-vision = "^3.4.0"
google-cloud-videointelligence = "^2.11.0"
streamlit = "^1.24.0"
vertexai = "^0.0.1"
google-cloud-aiplatform = "^1.25.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.flake8]
max-line-length = 88
exclude = [".venv", ".git", "__pycache__", "build", "dist"] 